- worked on hover change - added some styles
