+++ src/ReactDateTimeRangePicker.css
@@ -1,6 +1,7 @@
 .rdtrp-container {
-    display: flex;
+    display: inline-flex;
     position: relative;
+    font-size: 14px;
 }
 
 .rdtrp-container .calendar-popup {
