+++ src/components/CalendarControls/Right/CalendarControlRight.css
@@ -5,3 +5,7 @@
 .right::after {
     content: '>';
 }
+
+.month-name {
+    margin-right: 5px;
+}
