+++ src/components/CalendarControls/Right/Index.tsx
@@ -6,7 +6,8 @@ function CalendarControlRight(props: CalendarControlProps) {
   return (
     <div className="right-control-container">
       <div className="centre">
-        {props.monthName}/{props.year}
+        <span className="month-name">{props.monthName}</span>
+        <span className="year-name">{props.year}</span>
       </div>
       <div className="right" onClick={() => props.increaseMonth!()}></div>
     </div>
