+++ src/components/CalendarLayout/CalendarLayout.css
@@ -5,4 +5,9 @@
 .calendar-layout .left-control-container, .calendar-layout .right-control-container {
     display: flex;
     justify-content: space-evenly;
+    font-weight: bold;
+}
+
+.calendar-layout .left-calendar {
+    margin-right: 20px;
 }
