+++ src/components/Day/Day.css
@@ -11,6 +11,7 @@
     flex: 1;
     cursor: pointer;
     text-align: center;
+    padding: 5px 0;
 }
 
 .default-hover:hover {
